backend/requirements.txt
APScheduler==3.6.3
asciimatics==1.14.0
asgiref==3.5.0
atomicwrites==1.4.1
attrs==22.2.0
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.1
defusedxml==0.7.1
Django==3.2.13
django-debug-toolbar==3.2.4
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-extra-fields==3.4.0
Faker==12.0.1
flake8==4.0.1
flake8-broken-line==0.4.0
flake8-docstrings==1.6.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
future==0.18.2
gunicorn==20.0.4
idna==3.3
importlib-metadata==6.0.0
iniconfig==2.0.0
isort==5.10.1
isoweek==1.3.3
itypes==1.2.0
Jinja2==3.1.1
Markdown==3.4.1
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==9.0.0
oauthlib==3.2.0
packaging==23.0
pep8-naming==0.12.1
Pillow==9.1.0
pluggy==1.0.0
psycopg2-binary==2.8.6
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pydocstyle==6.3.0
pyfiglet==0.8.post1
pyflakes==2.4.0
PyJWT==2.3.0
pytest==6.2.5
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.20.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytils==0.4.1
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
pywin32==305
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.2.0
sorl-thumbnail==12.8.0
sqlparse==0.4.2
testfixtures==6.18.5
TimeConvert==3.0.13
toml==0.10.2
tornado==6.2
tzdata==2023.3
tzlocal==4.3
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
xlwt==1.3.0
zipp==3.14.0
